+++ mayan/apps/cabinets/apps.py
@@ -6,10 +6,11 @@ from mayan.apps.acls.permissions import (
     permission_acl_edit, permission_acl_view
 )
 from mayan.apps.common.apps import MayanAppConfig
-from mayan.apps.common.classes import ModelCopy, ModelQueryFields
+from mayan.apps.common.classes import ModelCopy
 from mayan.apps.common.menus import (
     menu_list_facet, menu_main, menu_multi_item, menu_object, menu_secondary
 )
+from mayan.apps.databases.classes import ModelQueryFields
 from mayan.apps.events.classes import EventModelRegistry, ModelEventType
 from mayan.apps.navigation.classes import SourceColumn
 from mayan.apps.rest_api.fields import DynamicSerializerField