+++ mayan/apps/documents/views/document_views.py
@@ -5,7 +5,7 @@ from django.contrib import messages
 from django.urls import reverse
 from django.utils.translation import ugettext_lazy as _, ungettext
 
-from mayan.apps.common.classes import ModelQueryFields
+from mayan.apps.databases.classes import ModelQueryFields
 from mayan.apps.views.generics import (
     MultipleObjectFormActionView, SingleObjectDetailView,
     SingleObjectEditView, SingleObjectListView