+++ mayan/apps/templating/widgets.py
@@ -2,7 +2,8 @@ from django import forms
 from django.apps import apps
 from django.contrib import admindocs
 from django.utils.translation import gettext as _
-from mayan.apps.common.classes import ModelAttribute
+
+from mayan.apps.databases.classes import ModelAttribute
 from mayan.apps.views.widgets import NamedMultiWidget
 
 from .classes import Template